requirements.txt
bacpypes==0.17.0
google-api-python-client==1.6.5
google-auth-httplib2==0.0.3
google-auth==1.4.1
cryptography==2.1.4
paho-mqtt==1.3.1
pyjwt==1.6.0
netifaces==0.10.6